@@ -2,18 +2,12 @@
 
 namespace Jsor\HalClient\HttpClient;
 
-use Psr\Http\Message\RequestInterface;
-use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\{RequestInterface, ResponseInterface};
 
 interface HttpClientInterface
 {
     /**
-     * Note, that this method must not throw exceptions but always return a
-     * response.
-     *
-     * @param RequestInterface $request
-     *
-     * @return ResponseInterface
+     * Note, that this method must not throw exceptions but always return a response.
      */
-    public function send(RequestInterface $request);
+    public function send(RequestInterface $request) : ResponseInterface;
 }
